working code

+++ src/main/java/com/vishav/rewards/domain/Transaction.java
@@ -11,17 +11,17 @@ import java.util.Date;
  */
 public class Transaction {
 	
-	private int customerId;
+	private long customerId;
 	
 	private double transactionAmount;
 	
 	private String transactionDate;
 
-	public int getCustomerId() {
+	public long getCustomerId() {
 		return customerId;
 	}
 
-	public void setCustomerId(int customerId) {
+	public void setCustomerId(long customerId) {
 		this.customerId = customerId;
 	}
 
@@ -40,7 +40,5 @@ public class Transaction {
 	public void setTransactionDate(String transactionDate) {
 		this.transactionDate = transactionDate;
 	}
-	
-	
 
 }
